requirements.txt
# Web framework
fastapi==0.135.3
uvicorn==0.43.0
python-multipart==0.0.22
slowapi==0.1.9

# PDF parsing
PyPDF2==3.0.1

# Embeddings (local, free)
sentence-transformers==3.4.1

# OpenAI (GPT-4 generation only)
openai==2.30.0

# LangChain (required by ragas)
langchain==1.2.15
langchain-openai==1.1.12
langchain-community==0.4.1

# Vector database
pinecone==8.1.2

# Reranking
cohere>=5.0,<6.0

# Evaluation
ragas==0.4.3
datasets==4.8.4
bert-score==0.3.13

# Utilities
python-dotenv==1.2.2
pandas==2.3.3
numpy>=1.26,<2.0
tiktoken==0.12.0

# BM25 for hybrid retrieval + ablation
rank-bm25==0.2.2

# Fuzzy skill matching
rapidfuzz==3.12.2

# Test resume PDF generation
fpdf2==2.8.5

# Kaggle dataset download
kaggle==1.7.4.5
